pyproject.toml
[project]
name = "jairocloud-groups-manager"
version = "0.1.0"
authors = [
  {name = "National Institute of Informatics"}
]
description = "A web client for JAIRO Cloud repository administrators to manage groups in the GakuNin Cloud Gateway."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.14"
dependencies = [
    "celery[redis]>=5.6.2",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-pydantic>=0.14.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg[binary]>=3.3.2",
    "flask-login>=0.6.3",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "requests>=2.32.5",
    "sqlalchemy-utils>=0.42.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest]
pythonpath = ["."]
addopts = [
    "--import-mode=importlib",
]

[tool.coverage.run]
branch = true
source = ["src/server"]

[tool.coverage.report]
omit = ["__init__.py", "tests/*"]
exclude_also = [
  "if t.TYPE_CHECKING:",
]


[tool.ruff.lint]
# refer to https://docs.astral.sh/ruff/settings/#lint
ignore = [
  "COM812",
  "INP001",
  "D203",
  "D213",
  "TRY400",
]
preview = true
select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/api/**.py" = ["TC001", "TC002", "TC003"]
"**/entities/**.py" = ["TC001", "TC002", "TC003"]
"*.pyi" = ["CPY001"]

[tool.ruff.lint.isort]
# refer to https://docs.astral.sh/ruff/settings/#lintisort
combine-as-imports = true
lines-after-imports = 2
lines-between-types = 1

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.analyze]
# refer to https://docs.astral.sh/ruff/settings/#analyze
detect-string-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
# refer to https://docs.astral.sh/ruff/settings/#format
preview = true
skip-magic-trailing-comma = false


[tool.pyright]
extraPaths = ["src/server"]
typeCheckingMode =  "standard"
